Refactored report file with builder pattern
