+++ src/pytest_bdd_report/templates/report_template.py
@@ -70,4 +70,3 @@ class ReportTemplateBuilder:
 
     def build(self) -> ReportTemplate:
         return self.report_template
-
